4.09

+++ SITE/delete.php
@@ -1,4 +1,11 @@
  <?php
+
+if (!isset($_COOKIE["username"]))
+{
+    header("Location: login.php");
+    die();
+}
+
 require('db.php');
 $id=$_REQUEST['id'];
 $stmt = $con->prepare("DELETE FROM knigi WHERE id=?");
